chore: move test in test_dataclasses
